+++ src/plone/versioncheck/formatter.py
@@ -1,145 +1,168 @@
 # -*- coding: utf-8 -*-
-from colorama import init as colorama_init
-from colorama import Fore
-from colorama import Style
+from __future__ import print_function
+from collections import OrderedDict
 from plone.versioncheck import analyser
-
-# color codes:
-
-# pkgname all bright
-# ------------------
-# white: no newer pypi, no newer cfgs
-# green:  newer pypi finals, no newer cfgs
-# blue: newer pypi prerelease only, no newer cfgs
-# red:  newer in both, pypi final and cfgs
-# cyan:  newer in both, pypi prerelease and cfgs
-# yellow newer in cfgs
-
-# all dots gray dimmed
-
-# version / location (w/o dots)
-# white current used
-# dimmed white unused and older
-# green pypi final newer
-# blue pypi prerelease newer
-# yellow cfgs newer
+from plone.versioncheck.utils import color_by_state
+from plone.versioncheck.utils import color_init
+from plone.versioncheck.utils import dots
+import json
+import sys
 
 
-def dots(value, number):
-    """ljust, but the dots only"""
-    return '.' * (number - len(value))
-
-
-def format_line(name,
-                pkg, pypi,
-                key, idx,
-                ver_maxlen, pkg_maxlen,
-                flavor='versions',
-                colored=True):
-    # prepare
-    line = Style.RESET_ALL if colored else ''
-
-    # before version
-    if idx == 0:
-        # format
-        if colored:
-            line += Style.BRIGHT
-            state = analyser.uptodate_analysis(pkg, pypi)
-            if 'cfg' in state and 'pypifinal' in state:
-                line += Fore.RED
-            elif 'cfg' in state and 'pypiprerelease' in state:
-                line += Fore.CYAN
-            elif 'cfg' in state:
-                line += Fore.YELLOW
-            elif 'pypifinal' in state:
-                line += Fore.GREEN
-            elif 'pypiprerelease' in state:
-                line += Fore.BLUE
-            else:
-                line += Fore.WHITE
-        line += name
-        if colored:
-            line += Style.DIM + Fore.WHITE
-        line += dots(name, pkg_maxlen)
-    else:
-        line += ' ' * (pkg_maxlen)
-    line += Style.RESET_ALL if colored else ''
-
-    # versions
-    color = ''
-    displayname = key
+def build_version(
+    name,
+    pkg,
+    pypi,
+    tracked,
+    key,
+    idx,
+    flavor='versions',
+):
+    record = {}
     if flavor == 'versions':
-        version = pkg[key]
-        if colored:
-            if idx == 0:
-                color = Style.BRIGHT + Fore.WHITE
-            elif analyser.is_cfgidx_newer(pkg, idx):
-                color = Fore.YELLOW
-            else:
-                color = Style.DIM + Fore.WHITE
-            pass
-    else:
-        if colored:
-            # find a color
-            color = Style.BRIGHT + Fore.BLUE
-            pass
-        displayname = key + ' (PyPI)'
-        version = pypi[key]
-
-    line += ' ' + color + version + dots(version, ver_maxlen)
-    line += ' ' + color + displayname
-    line += Style.RESET_ALL if colored else ''
-
-    return line
+        record['version'] = pkg[key] or "(unset)"
+        record['description'] = key
+        if idx == 0:
+            record['state'] = 'I' if tracked and tracked[1] else 'A'
+        elif analyser.is_cfgidx_newer(pkg, idx):
+            record['state'] = 'In'
+        else:
+            record['state'] = 'I'
+    else:  # pypi
+        if 'pre' in key:
+            record['state'] = 'P'
+        else:
+            record['state'] = 'U'
+        record['version'] = pypi[key]
+        record['description'] = key.capitalize()
+    return record
 
 
-def display(pkgsinfo, overrides_only=False, colored=True, limit=None):
-    color = ''
-    if colored:
-        colorama_init()
-        color += Style.BRIGHT
-        color += Fore.CYAN
-    print color + 'Check Versions'
-    if colored:
-        color = Style.DIM
-    print color + '--------------'
+def builder(pkgsinfo, newer_only=False, limit=None):
+    """build
+    - OrderedDict with pkgname as keys
+    - each entry an record:
+      - state: overall package state
+      - versions: list of dicts with
+        - version number
+        - state
+        - description
+    """
+    result = OrderedDict()
+    ver_maxlen = 0
     pkgs = pkgsinfo['pkgs']
     pypi = pkgsinfo.get('pypi', {})
-    for nidx, name in enumerate(sorted(pkgs)):
-        if limit and limit == nidx:
-            break
-        if overrides_only and len(pkgs[name]) < 2:
-            continue
-        current_pkg = pkgs[name]
+    tracked = pkgsinfo.get('tracking', {}).get('versions', {})
+    requ = pkgsinfo.get('tracking', {}).get('required_by', {})
+
+    names = sorted(set(tracked.keys()) | set(pkgs.keys()))
+
+    for nidx, name in enumerate(names):
+        current_pkg = pkgs.get(name, {})
+        record = dict()
+        versions = record['versions'] = list()
+        unpinned = False
+        required_by = requ.get(name, None)
+        if required_by:
+            record['required_by'] = required_by
+
+        # handle dev-eggs
+        devegg = False
+        current_tracked = tracked.get(name, None)
+        if current_tracked is not None:
+            ver_maxlen = max([ver_maxlen, len(current_tracked[0])])
+            if current_tracked[1]:
+                versions.append({
+                    'version': current_tracked[0],
+                    'state': 'D',
+                    'description': current_tracked[1],
+                })
+                devegg = True
+
+        # handle versions.cfg and inherited
         for idx, location in enumerate(current_pkg):
-            print \
-                format_line(
+            ver_maxlen = max([ver_maxlen, len(current_pkg.get(name, ''))])
+            versions.append(
+                build_version(
                     name,
                     current_pkg,
                     pypi.get(name, {}),
+                    current_tracked,
                     location,
                     idx,
-                    pkgsinfo['ver_maxlen'],
-                    pkgsinfo['pkg_maxlen'],
                     flavor='versions',
-                    colored=colored
+                )
+            )
+        if not devegg and current_tracked is not None and not len(versions):
+            ver_maxlen = max([ver_maxlen, len(current_tracked[0])])
+            versions.append({
+                'version': current_tracked[0],
+                'state': 'X',
+                'description': 'unpinned',
+            })
+            unpinned = True
+
+        if pypi.get(name, None) is not None:
+            current_pypi = pypi[name]
+            for label, version in current_pypi.items():
+                if version is None:
+                    continue
+                versions.append(
+                    build_version(
+                        name,
+                        current_pkg,
+                        current_pypi,
+                        current_tracked,
+                        label,
+                        idx,
+                        flavor='pypi',
+                    )
                 )
 
-        if pypi.get(name, None) is None:
+        pkgsinfo['ver_maxlen'] = ver_maxlen
+        states = analyser.uptodate_analysis(
+            current_pkg,
+            pypi.get(name, {}),
+        )
+        if devegg:
+            # dev always wins - not true!
+            record['state'] = 'D'
+        elif unpinned:
+            record['state'] = 'X'
+        elif 'pypifinal' in states:
+            record['state'] = 'U'
+        elif 'cfg' in states:
+            record['state'] = 'In'
+        elif 'pypifinal' in states:
+            record['state'] = 'P'
+        else:
+            record['state'] = 'A'
+
+        if newer_only and record['state'] == 'A':
             continue
-        current_pypi = pypi[name]
-        for label, version in current_pypi.items():
-            if version is None:
-                continue
-            print \
-                format_line(
-                    name,
-                    current_pkg,
-                    current_pypi,
-                    label,
-                    -1,
-                    pkgsinfo['ver_maxlen'],
-                    pkgsinfo['pkg_maxlen'],
-                    flavor='pypi',
-                    colored=colored
-                )
+
+        result[name] = record
+    return result
+
+
+def human(pkgsinfo, newer_only=False, limit=None):
+    color_init()
+    sys.stderr.write('\nReport for humans\n\n')
+    data = builder(pkgsinfo, newer_only=newer_only, limit=limit)
+    for name, record in data.items():
+        print(color_by_state(record['state']) + name)
+        for version in record['versions']:
+            print(
+                ' ' * 4 +
+                color_by_state(version['state']) +
+                version['version'] + ' ' +
+                dots(version['version'], pkgsinfo['ver_maxlen']-1) +
+                ' ' + color_by_state(version['state']) +
+                version['state'][0] + ' ' + version['description']
+            )
+
+
+def machine(pkgsinfo, newer_only=False, limit=None):
+    sys.stderr.write('\nReport for machines\n\n')
+    data = builder(pkgsinfo, newer_only=newer_only, limit=limit)
+    print(json.dumps(data, indent=4))